+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Refresh Affirm UI on PDP if quantity changes
+
 ## [0.0.7] - 2022-11-04
 
 ### Fixed